requirements.txt
streamlit==1.31.0
groq==0.4.2
sumolib==1.19.0
traci==1.19.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
python-dotenv==1.0.0